requirements.txt
folium==0.17.0
haversine==2.8.1
image==1.5.33
inflection==0.5.1
matplotlib-inline==0.1.7
numpy==2.1.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
pyarrow==17.0.0
streamlit==1.39.0
streamlit_folium==0.23.1